# services/fk_backend/conf/swagger-ui/swagger.json
{
  "components" : {
    "schemas" : {
...
      "library.dtos.AuthorPaginateDTO" : {
        "properties" : {
          "count" : {
            "format" : "int32",
            "type" : "integer"
          },
          "authors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/library.dtos.AuthorDTO"
            }
          }
        },
        "required" : [ "authors", "count" ]
      },
      "foundation.dtos.QueryParamSorterDTO" : {
        "properties" : {
          "sortName" : {
            "type" : "string"
          },
          "sortOrder" : {
            "type" : "string"
          },
          "tableName" : {
            "type" : "string"
          }
        },
        "required" : [ "tableName", "sortOrder", "sortName" ]
      },
      "foundation.dtos.QueryParamDTO" : {
        "properties" : {
          "drop" : {
            "nullable" : true,
            "format" : "int32",
            "type" : "integer"
          },
          "take" : {
            "nullable" : true,
            "format" : "int32",
            "type" : "integer"
          },
          "sorter" : {
            "nullable" : true,
            "$ref" : "#/components/schemas/foundation.dtos.QueryParamSorterDTO"
          },
          "filters" : {
            "nullable" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/foundation.dtos.QueryParamFilterDTO"
            }
          }
        }
      },
      "library.dtos.BookDTO" : {
        "properties" : {
          "bookStores" : {
            "nullable" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/library.dtos.BookStoreDTO"
            }
          },
          "language_id" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "nullable" : true,
            "format" : "int64",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "author_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "published_in" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "author_id", "title", "published_in", "language_id" ]
      },
      "library.dtos.BookStoreDTO" : {
        "properties" : {
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "name" ]
      },
      "foundation.dtos.QueryParamFilterDTO" : {
        "properties" : {
          "filterValue" : {
            "type" : "string"
          },
          "filterName" : {
            "type" : "string"
          },
          "filterComparator" : {
            "type" : "string"
          },
          "tableName" : {
            "type" : "string"
          }
        },
        "required" : [ "tableName", "filterName", "filterValue", "filterComparator" ]
      },
      "library.dtos.AuthorDTO" : {
        "properties" : {
          "books" : {
            "nullable" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/library.dtos.BookDTO"
            }
          },
          "date_of_birth" : {
            "nullable" : true,
            "format" : "int32",
            "type" : "integer"
          },
          "reviewed_at" : {
            "nullable" : true,
            "format" : "int32",
            "type" : "integer"
          },
          "created_at" : {
            "format" : "int32",
            "type" : "integer"
          },
          "last_name" : {
            "type" : "string"
          },
          "id" : {
            "nullable" : true,
            "format" : "int64",
            "type" : "integer"
          },
          "distinguished" : {
            "nullable" : true,
            "format" : "int32",
            "type" : "integer"
          },
          "first_name" : {
            "nullable" : true,
            "type" : "string"
          },
          "year_of_birth" : {
            "nullable" : true,
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "last_name", "created_at" ]
      }
    },
    "securitySchemes" : {
      "cookieAuth" : {
        "type" : "apiKey",
        "in" : "cookie",
        "name" : "JSESSIONID"
      }
    }
  },
  "openapi" : "3.0.0",
  "paths" : {
    "/api/library/author/delete/{id}" : {
      "post" : {
        "operationId" : "deleteAuthor",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "tags" : [ "library" ]
      }
    },
...
    "/api/library/author/add" : {
      "post" : {
        "operationId" : "addAuthor",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "schema" : {
            "$ref" : "#/components/schemas/library.dtos.AuthorDTO"
          }
        } ],
        "consumes" : [ "application/json" ],
        "tags" : [ "library" ]
      }
    },
...
    "/api/library/author/paginate" : {
      "post" : {
        "security" : [ {
          "cookieAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/foundation.dtos.QueryParamDTO"
              },
              "examples" : {
                "example1" : {
                  "summary" : "Select all authors with last_name Orwell,Coelho, which have published books in 1948,1945,1988 and return them sorted by last_name ascending",
                  "value" : {
                    "drop" : 0,
                    "take" : 10,
                    "sorter" : {
                      "sortName" : "last_name",
                      "sortOrder" : "asc",
                      "tableName" : "author"
                    },
                    "filters" : [ {
                      "tableName" : "author",
                      "filterName" : "last_name",
                      "filterValue" : "Orwell,Coelho",
                      "filterComparator" : "in"
                    }, {
                      "tableName" : "book",
                      "filterName" : "published_in",
                      "filterValue" : "1948,1945,1988",
                      "filterComparator" : "in"
                    } ]
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "schema" : {
              "$ref" : "#/components/schemas/library.dtos.AuthorPaginateDTO"
            },
            "description" : "success"
          },
          "500" : {
            "description" : "internal server error"
          }
        },
        "operationId" : "paginate",
        "tags" : [ "library" ]
      }
    },
...
  },
  "info" : {
    "description" : "Example Project for a Backend Server-Framework using Play/Slick",
    "title" : "Funkrusher Framework",
    "version" : "1.0-SNAPSHOT"
  },
...
}